修复图标输出一直为0，开启曲线绘制。

+++ Assets/Scripts/CoreLogic/TrajectoryData.cs
@@ -191,9 +191,9 @@ namespace TrajectoryData
             vec = traj.vec;
         }
 
-        public uint size()
+        public int size()
         {
-            return (uint)vec.Count;
+            return vec.Count;
         }
 
         public Trajectory append(Trajectory traj)
